listen to array changes
